Try to improve Boruvka algo (#10)
* fix segfault for Boruvka

* fix spla

* fix spla

* feat: first try

* fix tests

* small refactoring

* feat: prepare for experiment

* feat: optimize pt1

* feat: optimize pt2

* feat: optimize pt3

* feat: final optimization

* fix tests

* refactor

* fix test graph

---------

+++ src/lib/common/tree.hpp
@@ -5,8 +5,8 @@ namespace algos {
         int n;
         // parent[i] = -1 if i is root
         std::vector<int> parent;
-        float weight;
+        uint64_t weight;
 
-        Tree(int n, std::vector<int> parent, float w) : n(n), parent(std::move(parent)), weight(w) {}
+        Tree(int n, std::vector<int> parent, uint64_t w) : n(n), parent(std::move(parent)), weight(w) {}
     };
 }
